ecs-pytorch-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.0.0+cpu
torchvision==0.15.0+cpu
Pillow==10.0.1
numpy==1.24.3
python-multipart==0.0.6
boto3==1.34.0